+++ autoware_utils_tf2/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>autoware_utils</name>
+  <name>autoware_utils_tf2</name>
   <version>1.1.0</version>
-  <description>The autoware_utils package</description>
+  <description>The autoware_utils_tf2 package</description>
   <maintainer email="egon.kang@autocore.ai">Jian Kang</maintainer>
   <maintainer email="ryohsuke.mitsudome@tier4.jp">Ryohsuke Mitsudome</maintainer>
   <maintainer email="esteve.fernandez@tier4.jp">Esteve Fernandez</maintainer>
@@ -13,24 +13,9 @@
   <buildtool_depend>ament_cmake_auto</buildtool_depend>
   <buildtool_depend>autoware_cmake</buildtool_depend>
 
-  <depend>autoware_internal_debug_msgs</depend>
-  <depend>autoware_internal_msgs</depend>
-  <depend>autoware_internal_planning_msgs</depend>
-  <depend>autoware_perception_msgs</depend>
-  <depend>autoware_planning_msgs</depend>
-  <depend>autoware_utils_tf2</depend>
-  <depend>autoware_vehicle_msgs</depend>
-  <depend>builtin_interfaces</depend>
-  <depend>diagnostic_msgs</depend>
   <depend>geometry_msgs</depend>
-  <depend>libboost-system-dev</depend>
-  <depend>logging_demo</depend>
   <depend>rclcpp</depend>
-  <depend>tf2</depend>
-  <depend>tf2_eigen</depend>
-  <depend>tf2_geometry_msgs</depend>
-  <depend>unique_identifier_msgs</depend>
-  <depend>visualization_msgs</depend>
+  <depend>tf2_ros</depend>
 
   <test_depend>ament_cmake_ros</test_depend>
   <test_depend>ament_lint_auto</test_depend>
